concord-server: remove imports from ProcessEntry
The only place where we read `Imports` is when a process fork is created.
Removing it from `ProcessEntry` saves a bit of time by not deserializing the data.
Also helps us to avoid issues when an unknown import type is present in the `process_queue` data.
Before this change it would lead to a deserialization error when trying to get a list of processes.