runtime-v2: actually working checkpoints
Update the runner v2 to actually support checkpoints.
The previous implementation expected a checkpoint to have an eventRef associated with it.
Which was the case in v1, but not in v2. In v2 we can directly feed a ProcessSnapshot instance into the runtime.

Update the runner v2 and add some tests.