concord-agent: option to keep workDir after the process ends (#120)
A new configuration option:
```
concord-agent {
  runner {
    persistentWorkDir = "/path/to/dir"
  }
}
```

If specified, Agent will copy all files from the process' working directory
into a subdirectory of `persistentWorkDir` after the process ends (regardless
of the status).

This feature is planned to be used in testcontainers-concord to allow test authors
to inspect the process ${workDir}.

All copied files will be readable by all users (i.e. it sets `0644`
permissions for files and `755` for directories). Should *not* be used in production.