runtime-v2: use eventRefs in checkpoints (#231)
When creating a checkpoint, use marker files instead of
`executionMode(CHECKPOINT_RESTORE)`.

Normally, when creating a fork we use the last known state of the parent
process. When restoring the parent process from a checkpoint, we end up
with the process state that is marked as `executionMode(CHECKPOINT_RESTORE)`.
If such state ends up in a fork then the runtime thinks it's a
checkpoint restoration. The fork ends up running the same thing as the
parent process.

This PR fixes it by switching to "marker files" instead of
`executionMode`. When a fork is created, the server removes such marker
files and the fork executes as it should - from the specified `entryPoint`
with the state inherited from the parent process.