runtime-v2: use the dependencies classloader for checkpoints
Explicitly specify the classloader when deserializing checkpoints.

Before the change, checkpoints used the default classloader of the
runner which doesn't contain classes from `dependencies`.

While it is still not recommended to use 3rd party classes as flow
variables, it might be okay for situations when both the server
and the runtime have the same 3rd party JARs on the classpath.