concord-server: truncate createdAt nanos (#736)
Postgres stores timestamps with microsecond precision. When the
createdAt portion of a process key is inserted, it might be either
rounded up by the PG JDBC driver or truncated by JOOQ. I.e. `createdAt`
may be saved with a +/- 1 microsecond difference.

The PR forces all process key to be truncated to nearest microsecond
before inserting into the DB.

---------

Co-authored-by: brig <ybrigo@gmail.com>