concord-server: do not invalidate sessions in onFailedLogin
This effectively reverts #859. The change broke redirect after login
when the authentication times out (e.g. OIDC login). Unfortunaly, pax4j
stores the "from" parameter in the session so we need it alive when we
redirect the user between the app and the auth provider.

We might want to use a separate cookie or have a way to differentiate between
onLoginFailure reasons when we make a decision to invalidate the
session.