docker: configure safe.directory for git 2.35+
Git 2.35+ introduced safe.directory feature, i.e. fail if .git belongs to a user different that the current.
It is an issue for our integration tests where we share .git between the host and the containers.
This change effectively disables the safe.directory feature by configuring safe.directory=* at user-level in the base Docker image.