concord-server: split ProcessSecurityContext
Before this change ProcessSecurityContext was responsible for both
runAs (generic "run as this user ID" method) and runAsCurrentUser (i.e.
"run as the current user of the specified process ID").

This change separates those concerns. The new UserSecurityContext can be
used in situations where there is no notion of processes at all (i.e.
testing or re-use in other projects).